// apps/web/lib/api/generated/models/publicUserResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * llame API
 * llame auth and domain API
 * OpenAPI spec version: 0.1
 */

export interface PublicUserResponse {
  id: string;
  /** @nullable */
  name: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  emailVerified: string | null;
  /** @nullable */
  image: string | null;
}